requirements.txt
tensorflow
scikit-learn
matplotlib
opencv-python-headless
streamlit
Pillow
kaggle